+++ backend/data/collection/transformation/database/createSchema.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS Transformed.Article (
     brand TEXT,
     price NUMERIC(10,2),
     size TEXT,
-    packagesize REAL,
+    packagesize INT,
     amount REAL,
     unit TEXT,
     url TEXT
